Get Concepts Catalog

This function returns a data frame of available concept IDs and their names for a specified table in an OMOP CDM database.

Arguments

table

The name of the table for which to retrieve the concepts catalog.

Value

A data frame containing concept_id and concept_name columns representing available concepts in the specified table of the OMOP CDM database.