Construct the Call for Table Assignment

This function assists in constructing the call that will be sent from the DataSHIELD client to the server to perform the assignment of a table from an OMOP CDM database based on the parameters specified by the user.

Usage

getTableCall(
  resource,
  table,
  conceptFilter = NULL,
  columnFilter = NULL,
  personFilter = NULL,
  mergeColumn = NULL,
  dropNA = FALSE,
  wideLongitudinal = FALSE
)

Arguments

resource

The resource symbol representing the connection to the OMOP CDM database.

table

A string specifying the name of the table to be assigned.

conceptFilter

A numeric vector specifying concept IDs to filter the table by.

columnFilter

A string vector specifying column names to filter (select) in the table.

personFilter

A string specifying the symbol in the environment of a table from which to obtain person IDs.

mergeColumn

A string specifying the column name for merging operations with other tables. Defaults to "person_id" if not specified.

dropNA

A boolean indicating whether to drop empty columns. Defaults to FALSE.

wideLongitudinal

A boolean indicating whether to reshape the table to a wide format with numerically suffixed columns if it contains longitudinal data. Defaults to FALSE.

Value

A string representing the call to be executed on the DataSHIELD server for table assignment.