OMOP CDM Database class

This class represents a connection to an OMOP CDM database, providing various methods for the user to interact with the database. It allows for the retrieval of tables as well as fetching catalogs of available data within the database.

Methods

Public methods

OMOPCDMDatabase$new()

OMOPCDMDatabase$assignResource()

OMOPCDMDatabase$removeResource()

OMOPCDMDatabase$get()

OMOPCDMDatabase$tables()

OMOPCDMDatabase$columns()

OMOPCDMDatabase$concepts()

OMOPCDMDatabase$checkConnection()

OMOPCDMDatabase$checkPrivacyControlLevel()

OMOPCDMDatabase$clone()

Method new()

This method initializes a new instance of the OMOPCDMDatabase class. It sets up the connection to the OMOP CDM database given the provided DataSHIELD connection object and the OMOP CDM database resource identifier.

Usage

OMOPCDMDatabase$new(connections, resource)

Arguments

connections

Connection object to the DataSHIELD server.

resource

Either an identifier or a named list of identifiers for the specific resource(s) within the DataSHIELD server(s). If a named list, the name of each resource identifier should correspond to the server name in the connections.

Returns

A new instance of the OMOPCDMDatabase class.

Method assignResource()

Usage

OMOPCDMDatabase$assignResource(symbol)

Method removeResource()

Usage

OMOPCDMDatabase$removeResource()

Method get()

Usage

OMOPCDMDatabase$get(
  table,
  symbol = NULL,
  conceptFilter = NULL,
  columnFilter = NULL,
  personFilter = NULL,
  mergeColumn = NULL,
  dropNA = FALSE
)

Method tables()

Usage

OMOPCDMDatabase$tables()

Method columns()

Usage

OMOPCDMDatabase$columns(table, dropNA = FALSE)

Method concepts()

Usage

OMOPCDMDatabase$concepts(table)

Method checkConnection()

Usage

OMOPCDMDatabase$checkConnection()

Method checkPrivacyControlLevel()

Usage

OMOPCDMDatabase$checkPrivacyControlLevel()

Method clone()

The objects of this class are cloneable with this method.

Usage

OMOPCDMDatabase$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.