Get Table from OMOP CDM Database

This function assigns a specified table by name (string) from an OMOP CDM database to the DataSHIELD environment. The user can specify the symbol (string) for the table assignment in the DataSHIELD environment, a vector of concept IDs (numeric) to filter if dealing with a table categorized by concept IDs, a vector of column names (strings) of the table to filter (i.e., select), the name of a symbol (string) in the environment of a table from which to obtain its person IDs (from the unique values of their person_id), and the name of the column (string) by which to perform merging operations with other database tables. If not specified, the "person_id" column is assumed to be the merging column. The user is also provided with the option to drop empty columns.

Arguments

table

A string specifying the name of the table to be assigned.

symbol

A string specifying the symbol for the table assignment in the DataSHIELD environment.

conceptFilter

A numeric vector specifying concept IDs to filter the table by.

columnFilter

A string vector specifying column names to filter (select) in the table.

personFilter

A string specifying the symbol in the environment of a table from which to obtain person IDs.

mergeColumn

A string specifying the column name for merging operations with other tables. Defaults to "person_id" if not specified.

dropNA

A boolean indicating whether to drop empty columns. Defaults to FALSE.